방 관련 유저 연결 추가

+++ src/main/java/com/inha/endgame/dto/request/AddUserRequest.java
@@ -27,5 +27,5 @@ public class AddUserRequest implements ClientRequest {
 	@Schema(description = "ADD_USER", defaultValue = "ADD_USER")
 	private RequestType type;
 	private String userName;
-	private int roomId;
+	private long roomId;
 }
